Moved imgui-property related utility functions to get/restore property values of a object from the node_config module to type_editor
